Extracted Page as a component
