+++ src/app.js
@@ -5,8 +5,8 @@ import classy from 'classwrap';
 import { About } from './About';
 import { Contact, ContactModule } from './Contact';
 import { Footer } from './Footer';
-import { Loader } from './Loader';
 import { Navigation } from './Navigation';
+import { Page } from './Page';
 import { Posts, PostsModule } from './Posts';
 import { Profile } from './Profile';
 import { Skills, SkillsModule } from './Skills';
@@ -55,40 +55,3 @@ export const Actions = app({
       <Footer />
     </div>
 });
-
-const Page = (props) => {
-  const state = Actions.getState();
-  if (state.router.route !== props.route) return null;
-
-  const moduleState = state[props.module];
-  if (!moduleState) throw new Error(`Invalid module state for: ${props.module}`);
-
-  const moduleActions = Actions[props.module];
-  if (!moduleActions) throw new Error(`Invalid module actions for: ${props.module}`);
-
-  // NOTE: We don't need to set `$resolved` to false on element remove, because
-  // the data only has to be loaded once throughout the lifecycle of the application
-  // onremove={() => moduleState.$resolved = false}
-  if (props.resolve && !moduleState.$resolved) {
-    if (!moduleState.$pending) {
-      moduleState.$pending = true;
-      props.resolve()
-        .then(() => {
-          moduleState.$resolved = true;
-          moduleState.$pending = false;
-        })
-        .catch(() => {
-          moduleState.$pending = false;
-        })
-      ;
-    }
-    return <Loader key={props.module} />;
-  }
-
-  return <props.view key={props.module} state={moduleState} actions={moduleActions} onremove={fadeOutPage} />;
-};
-
-const fadeOutPage = (el) => (remove) => {
-  el.classList.add('page-fade-out');
-  setTimeout(remove, 1000 * animationDuration);
-};
